Fix humidity type parsing

+++ models/weather.go
@@ -7,6 +7,6 @@ import (
 type WeatherData struct {
 	ID          uint      `gorm:"primaryKey" json:"id"`
 	Date        time.Time `gorm:"uniqueIndex" json:"date"`
-	Humidity    int       `json:"humidity"`
+	Humidity    int64     `json:"humidity"`
 	Temperature float64   `json:"temperature"`
 }
